basic

+++ src/main/java/br/byteshop/ecommerce/bs_api/model/StatusPedido.java
@@ -5,11 +5,10 @@ import jakarta.persistence.*;
 @Entity
 @Table(name = "tb03_status_pedido", schema = "loja")
 public class StatusPedido {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_status_pedido", nullable = false)
-    private Long id;
+    private Integer id;
 
     @Column(name = "descricao", nullable = false, unique = true, length = 30)
     private String descricao;
@@ -22,11 +21,11 @@ public class StatusPedido {
         this.descricao = descricao;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
